TST: Reorganize btf tests into own directory.
